+++ miranda/plugins/tabsrmm/tabctrl.c
@@ -478,9 +478,9 @@ static void DrawItem(struct TabControlData *tabdat, HDC dc, RECT *rcItem, int nH
 			}
 #if defined(_UNICODE)
 			if (tabdat->m_skinning == FALSE || myGlobals.m_TabAppearance & TCF_NOSKINNING)
-				DrawText(dc, dat->newtitle, _tcslen(dat->newtitle), rcItem, dwTextFlags);
+				DrawText(dc, dat->newtitle, (int)(_tcslen(dat->newtitle)), rcItem, dwTextFlags);
 			else
-				pfnDrawThemeText(dwStyle & TCS_BUTTONS ? tabdat->hThemeButton : tabdat->hTheme, dc, 1, nHint & HINT_ACTIVE_ITEM ? 3 : (nHint & HINT_HOTTRACK ? 2 : 1), dat->newtitle, _tcslen(dat->newtitle), dwTextFlags, 0, rcItem);
+				pfnDrawThemeText(dwStyle & TCS_BUTTONS ? tabdat->hThemeButton : tabdat->hTheme, dc, 1, nHint & HINT_ACTIVE_ITEM ? 3 : (nHint & HINT_HOTTRACK ? 2 : 1), dat->newtitle, (int)(_tcslen(dat->newtitle)), dwTextFlags, 0, rcItem);
 #else
 			DrawText(dc, dat->newtitle, _tcslen(dat->newtitle), rcItem, dwTextFlags);
 #endif
